+++ packages/core/avatar-group/src/styled/MoreIndicator.js
@@ -1,6 +1,7 @@
 // @flow
+
 import styled from 'styled-components';
-import { colors, themed } from '@atlaskit/theme';
+import { colors, themed, withTheme } from '@atlaskit/theme';
 import {
   getBorderRadius,
   getInnerStyles,
@@ -22,11 +23,11 @@ const EXCESS_INDICATOR_FONT_SIZE: {
 const getBorderWidth = p =>
   p.isFocus && !p.isActive ? `${BORDER_WIDTH[p.size]}px` : 0;
 
-export const Outer = styled.button`
+export const Outer = withTheme(styled.button`
   ${getInnerStyles} background: 0;
-`;
+`);
 
-export const Inner = styled.span`
+export const Inner = withTheme(styled.span`
   background-color: ${themed({ light: colors.N40, dark: colors.DN70 })};
   border-radius: ${getBorderRadius};
   align-items: center;
@@ -39,4 +40,4 @@ export const Inner = styled.span`
   font-size: ${props => EXCESS_INDICATOR_FONT_SIZE[props.size]}px;
   justify-content: center;
   transition: box-shadow 200ms;
-`;
+`);