+++ packages/core/avatar/src/index.js
@@ -5,11 +5,12 @@ export { default as AvatarItem } from './components/AvatarItem';
 export { default as Presence } from './components/Presence';
 export { default as Status } from './components/Status';
 export { default as Skeleton } from './components/Skeleton';
+export * from './theme';
+export * from './theme/item';
 
 // The below are exposed for use by avatarGroup
 export { AVATAR_SIZES, BORDER_WIDTH } from './styled/constants';
 export type { AvatarClickType, AvatarPropTypes, SizeType } from './types';
 export { withPseudoState } from './hoc';
-
 export { getProps } from './helpers';
 export { getBorderRadius, getInnerStyles } from './styled/utils';