+++ packages/core/tree/src/components/Tree/Tree.js
@@ -10,7 +10,7 @@ import {
   type DraggableProvided,
   type DraggableStateSnapshot,
   type DroppableProvided,
-} from 'react-beautiful-dnd';
+} from 'react-beautiful-dnd-next';
 import { getBox } from 'css-box-model';
 import { calculateFinalDropPositions } from './Tree-utils';
 import type { Props, State, DragState } from './Tree-types';
@@ -18,8 +18,12 @@ import { noop } from '../../utils/handy';
 import { flattenTree } from '../../utils/tree';
 import type { FlattenedItem, ItemId, Path } from '../../types';
 import TreeItem from '../TreeItem';
-import { type DragActionType } from '../TreeItem/TreeItem-types';
-import { getDestinationPath } from '../../utils/flat-tree';
+import {
+  getDestinationPath,
+  getItemById,
+  getIndexById,
+} from '../../utils/flat-tree';
+import DelayedFunction from '../../utils/delayed-function';
 
 export default class Tree extends Component<Props, State> {
   static defaultProps = {
@@ -31,20 +35,21 @@ export default class Tree extends Component<Props, State> {
     renderItem: noop,
     offsetPerLevel: 35,
     isDragEnabled: false,
+    isNestingEnabled: false,
   };
 
   state = {
     flattenedTree: [],
   };
 
-  // Action caused dragging
-  dragAction: DragActionType = null;
   // State of dragging. Null when resting
   dragState: ?DragState = null;
   // HTMLElement for each rendered item
   itemsElement: { [ItemId]: ?HTMLElement } = {};
   // HTMLElement of the container element
   containerElement: ?HTMLElement;
+
+  expandTimer = new DelayedFunction(500);
 
   static getDerivedStateFromProps(props: Props, state: State) {
     return {
@@ -59,6 +64,7 @@ export default class Tree extends Component<Props, State> {
       draggedItemId: result.draggableId,
       source: result.source,
       destination: result.source,
+      mode: result.mode,
     };
     if (onDragStart) {
       onDragStart(result.draggableId);
@@ -66,22 +72,36 @@ export default class Tree extends Component<Props, State> {
   };
 
   onDragUpdate = (update: DragUpdate) => {
+    const { onExpand } = this.props;
+    const { flattenedTree } = this.state;
     if (!this.dragState) {
       return;
+    }
+
+    this.expandTimer.stop();
+    if (update.combine) {
+      const { draggableId } = update.combine;
+      const item: ?FlattenedItem = getItemById(flattenedTree, draggableId);
+      if (item && this.isExpandable(item)) {
+        this.expandTimer.start(() => onExpand(draggableId, item.path));
+      }
     }
     this.dragState = {
       ...this.dragState,
       destination: update.destination,
+      combine: update.combine,
     };
   };
 
   onDragEnd = (result: DropResult) => {
     const { onDragEnd, tree } = this.props;
     const { flattenedTree } = this.state;
+    this.expandTimer.stop();
     const finalDragState: DragState = {
       ...this.dragState,
       source: result.source,
       destination: result.destination,
+      combine: result.combine,
     };
     const { sourcePosition, destinationPosition } = calculateFinalDropPositions(
       tree,
@@ -90,10 +110,6 @@ export default class Tree extends Component<Props, State> {
     );
     onDragEnd(sourcePosition, destinationPosition);
     this.dragState = null;
-  };
-
-  onDragAction = (actionType: DragActionType) => {
-    this.dragAction = actionType;
   };
 
   onPointerMove = () => {
@@ -114,17 +130,35 @@ export default class Tree extends Component<Props, State> {
     if (
       this.dragState &&
       this.dragState.draggedItemId === flatItem.item.id &&
-      this.dragState.destination
+      (this.dragState.destination || this.dragState.combine)
     ) {
-      // We only change the if it's being dragged by keyboard or just dropped
-      if (this.dragAction === 'key' || snapshot.isDropAnimating) {
-        const { source, destination, horizontalLevel } = this.dragState;
-        return getDestinationPath(
-          flattenedTree,
-          source.index,
-          destination.index,
-          horizontalLevel,
-        );
+      const {
+        source,
+        destination,
+        combine,
+        horizontalLevel,
+        mode,
+      } = this.dragState;
+      // We only update the path when it's dragged by keyboard or drop is animated
+      if (mode === 'SNAP' || snapshot.isDropAnimating) {
+        if (destination) {
+          // Between two items
+          return getDestinationPath(
+            flattenedTree,
+            source.index,
+            destination.index,
+            horizontalLevel,
+          );
+        }
+        if (combine) {
+          // Hover on other item while dragging
+          return getDestinationPath(
+            flattenedTree,
+            source.index,
+            getIndexById(flattenedTree, combine.draggableId),
+            horizontalLevel,
+          );
+        }
       }
     }
     return flatItem.path;
@@ -132,6 +166,9 @@ export default class Tree extends Component<Props, State> {
 
   isDraggable = (item: FlattenedItem): boolean =>
     this.props.isDragEnabled && !item.item.isExpanded;
+
+  isExpandable = (item: FlattenedItem): boolean =>
+    !!item.item.hasChildren && !item.item.isExpanded;
 
   getDroppedLevel = (): ?number => {
     const { offsetPerLevel } = this.props;
@@ -190,7 +227,6 @@ export default class Tree extends Component<Props, State> {
               path={currentPath}
               onExpand={onExpand}
               onCollapse={onCollapse}
-              onDragAction={this.onDragAction}
               renderItem={renderItem}
               provided={provided}
               snapshot={snapshot}
@@ -204,6 +240,7 @@ export default class Tree extends Component<Props, State> {
   };
 
   render() {
+    const { isNestingEnabled } = this.props;
     const renderedItems = this.renderItems();
 
     return (
@@ -212,7 +249,7 @@ export default class Tree extends Component<Props, State> {
         onDragEnd={this.onDragEnd}
         onDragUpdate={this.onDragUpdate}
       >
-        <Droppable droppableId="list">
+        <Droppable droppableId="list" isCombineEnabled={isNestingEnabled}>
           {(provided: DroppableProvided) => {
             const finalProvided: DroppableProvided = this.patchDroppableProvided(
               provided,
@@ -220,6 +257,7 @@ export default class Tree extends Component<Props, State> {
             return (
               <div
                 ref={finalProvided.innerRef}
+                style={{ pointerEvents: 'auto' }}
                 onTouchMove={this.onPointerMove}
                 onMouseMove={this.onPointerMove}
                 {...finalProvided.droppableProps}