+++ packages/editor/editor-common/src/schema/nodes/doc.ts
@@ -42,7 +42,7 @@ import { ActionDefinition as Action } from '../marks/action';
 
 /**
  * @name top_level_node
- * @minItems 1
+ * @minItems 0
  */
 export type TopLevel = Array<
   | Panel