+++ packages/editor/editor-core/src/plugins/table/utils/decoration.tsx
@@ -1,22 +1,40 @@
-import { EditorState } from 'prosemirror-state';
 import { Node as PmNode } from 'prosemirror-model';
 import { Decoration, DecorationSet } from 'prosemirror-view';
+import { TableDecorations } from '../types';
+import { TableCssClassName as ClassName } from '../types';
 
-export const createHoverDecorationSet = (
+export const createHoverDecoration = (
   cells: { pos: number; node: PmNode }[],
-  state: EditorState,
   danger?: boolean,
-): DecorationSet => {
+): Decoration[] => {
   const deco = cells.map(cell => {
-    const classes = ['hoveredCell'];
+    const classes = [ClassName.HOVERED_CELL];
     if (danger) {
       classes.push('danger');
     }
 
-    return Decoration.node(cell.pos, cell.pos + cell.node.nodeSize, {
-      class: classes.join(' '),
-    });
+    return Decoration.node(
+      cell.pos,
+      cell.pos + cell.node.nodeSize,
+      {
+        class: classes.join(' '),
+      },
+      { id: TableDecorations.CONTROLS_HOVER },
+    );
   });
 
-  return DecorationSet.create(state.doc, deco);
+  return deco;
 };
+
+export const findHoverDecoration = (
+  decorationSet: DecorationSet,
+): Decoration[] =>
+  decorationSet
+    .find()
+    .reduce(
+      (decorationArr: Decoration[], deco: Decoration) =>
+        deco.spec.id === TableDecorations.CONTROLS_HOVER
+          ? decorationArr.concat(deco)
+          : decorationArr,
+      [],
+    );