+++ packages/media/media-ui/CHANGELOG.md
@@ -1,5 +1,8 @@
 # @atlaskit/media-ui
 
+## 6.0.2
+- [patch] workaround to handle string size comming from /collection/items [8e99323](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/8e99323)
+
 ## 6.0.1
 - [patch] Updated dependencies [65c6514](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/65c6514)
   - @atlaskit/docs@5.0.8